Reorganize utils/ to align with Justfile run phases (pre/midrun/post)
Move mid-run scripts (those that run between CAIRO runs and consume
earlier run outputs) into a new utils/midrun/ directory:
- copy_calibrated_tariff_from_run.py (from utils/pre/)
- create_seasonal_discount_tariff.py (from utils/pre/)
- compute_subclass_rr.py (from utils/post/)
- compute_seasonal_discount_inputs.py (from utils/post/)
- latest_run_output.sh (from utils/runtime/, now removed)

Reorder utils/Justfile sections to match: PRE → MIDRUN.
Update all Python imports and path references.
