+++ tests/test_copy_calibrated_tariff_from_run.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from utils.pre.copy_calibrated_tariff_from_run import (
+from utils.midrun.copy_calibrated_tariff_from_run import (
     copy_calibrated_tariff_from_run_dir,
     _tariff_keys_and_payloads,
 )
